gateway/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "gtc-gateway"
version = "0.1.0"
description = "Single-port ASGI gateway for The Great Trading Company MCP servers"
requires-python = ">=3.11"
dependencies = [
    "uvicorn[standard]>=0.27",
    "starlette>=0.36",
    "python-dotenv>=1.0",
    # The 6 LOB packages are installed by the Dockerfile from local directories.
    # For local dev: pip install -e ../sf-mcp-app -e ../snow-mcp-app ... then pip install -e .
]

[project.scripts]
gtc-gateway = "gateway.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gateway*"]
